pyproject.toml
[project]
name = "online-judge-api"
version = "0.1.0"
description = "Using LangChain to create a online judge system"
readme = "README.md"
authors = [
    { name = "yanjiunliu", email = "ggyy831214@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.5",
    "django-filter>=25.1",
    "djangorestframework==3.16.1",
    "drf-spectacular>=0.28.0",
    "markdown>=3.8.2",
    "psycopg2>=2.9.10",
    "pygments>=2.19.2",
]
